feat: Доработка модалок. Добавил тип вызываемой модалки в стейт редакса и доработал лейаут модалок, чтобы показывалась вызываемая модалка.
