+++ src/constants/index.ts
@@ -1,0 +1,3 @@
+export enum ModalsType {
+  authModal = 'Auth modal',
+}
