+++ src/store/sagas/index.ts
@@ -1,8 +1,11 @@
 import { put, takeEvery } from 'redux-saga/effects';
+import { PayloadAction } from '@reduxjs/toolkit';
 
 import { showModal, closeModal } from '../slices/modalSlice';
-function* visibleModal() {
-  yield put(showModal());
+import { ModalsType } from '../../constants';
+
+function* visibleModal(action: PayloadAction<ModalsType>) {
+  yield put(showModal(action.payload));
 }
 function* hideModal() {
   yield put(closeModal());
