+++ src/store/slices/modalSlice.ts
@@ -1,7 +1,9 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { IModalState } from '../../types/modal';
+import { ModalsType } from '../../constants';
 
 const initialState: IModalState = {
+  modalComponent: null,
   isModalVisible: false,
 };
 
@@ -9,7 +11,8 @@ const modalSlice = createSlice({
   name: 'modal',
   initialState,
   reducers: {
-    showModal: (state) => {
+    showModal: (state, action: PayloadAction<ModalsType>) => {
+      state.modalComponent = action.payload;
       state.isModalVisible = true;
     },
     closeModal: (state) => {
