+++ src/types/modal.ts
@@ -1,3 +1,6 @@
+import { ModalsType } from '../constants';
+
 export interface IModalState {
+  modalComponent: ModalsType | null;
   isModalVisible: boolean;
 }
